Rename and re-organize in folders
