+++ public/components/datasources/components/__tests__/manage_data_connections_description.test.tsx
@@ -7,7 +7,7 @@ import { configure, mount } from 'enzyme';
 import Adapter from 'enzyme-adapter-react-16';
 import { waitFor } from '@testing-library/react';
 import React from 'react';
-import { DataConnectionsDescription } from '../manage_data_connections_description';
+import { DataConnectionsDescription } from '../manage/manage_data_connections_description';
 
 describe('Manage Data Connections Description test', () => {
   configure({ adapter: new Adapter() });
