+++ public/components/datasources/components/manage/connection_details.tsx
@@ -14,10 +14,10 @@ import {
 import React, { useState } from 'react';
 import { EuiPanel } from '@elastic/eui';
 import { ConnectionManagementCallout } from './connection_management_callout';
-import { coreRefs } from '../../../../public/framework/core_refs';
-import { DATACONNECTIONS_BASE } from '../../../../common/constants/shared';
-import { SaveOrCancel } from './save_or_cancel';
-import { ConnectionConfiguration } from './connection_configuration';
+import { coreRefs } from '../../../../framework/core_refs';
+import { DATACONNECTIONS_BASE } from '../../../../../common/constants/shared';
+import { SaveOrCancel } from '../save_or_cancel';
+import { ConnectionConfiguration } from '../new/connection_configuration';
 
 interface ConnectionDetailProps {
   dataConnection: string;
