+++ public/components/datasources/components/manage/query_permissions.tsx
@@ -17,8 +17,8 @@ import {
   OPENSEARCH_DOCUMENTATION_URL,
   QUERY_ALL,
   QUERY_RESTRICTED,
-} from '../../../../common/constants/data_connections';
-import { PermissionsConfigurationProps } from '../../../../common/types/data_connections';
+} from '../../../../../common/constants/data_connections';
+import { PermissionsConfigurationProps } from '../../../../../common/types/data_connections';
 
 export const QueryPermissionsConfiguration = (props: PermissionsConfigurationProps) => {
   const { roles, selectedRoles, setSelectedRoles } = props;
