+++ public/components/datasources/home.tsx
@@ -6,10 +6,10 @@
 import React from 'react';
 import { HashRouter, Route, RouteComponentProps, Switch } from 'react-router-dom';
 import { ChromeBreadcrumb, ChromeStart, HttpStart } from '../../../../../src/core/public';
-import { DataConnection } from './components/data_connection';
+import { DataConnection } from './components/manage/data_connection';
 import { ManageDataConnectionsTable } from './components/manage_data_connections_table';
-import { NewDatasource } from './components/new_datasource';
-import { Configure } from './components/configure_datasource';
+import { NewDatasource } from './components/new/new_datasource';
+import { Configure } from './components/new/configure_datasource';
 
 export interface HomeProps extends RouteComponentProps {
   pplService: any;
